pyproject.toml
[tool.poetry]
name = "valorant"
version = "0.1.0"
description = ""
authors = ["Jason Ren"]

[tool.poetry.dependencies]
python = "~3.9"
requests = "^2.26.0"
notebook = "^6.4.3"
numpy = "^1.21.2"
torch = "^1.9.0"
lxml = "^4.6.3"
black = "^21.7b0"
airtable-python-wrapper = "^0.15.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
